Added eu cookies notification plugin
This plugin is disabled by default. To activate you only need to rename the
file cookies-eu.js.disabled to cookies-eu.js and build.

This plugin doesn't integrate with angular, it's external plugin that makes
all its own with jquery.

Texts are in spanish and should be reviewed. I tried to use i18next for
translations but translations aren't ready in the initialization. Maybe we
should wait a second to show the notification and then the translations
will be there.